+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/enchantment/AirBorn.java
@@ -1,0 +1,49 @@
+package insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantment;
+
+import insane96mcp.iguanatweaksexpanded.module.experience.enchantments.NewEnchantmentsFeature;
+import insane96mcp.insanelib.world.enchantments.IEnchantmentTooltip;
+import net.minecraft.ChatFormatting;
+import net.minecraft.network.chat.Component;
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.enchantment.DiggingEnchantment;
+import net.minecraft.world.item.enchantment.Enchantment;
+import net.minecraft.world.item.enchantment.EnchantmentCategory;
+import net.minecraft.world.level.block.state.BlockState;
+
+public class AirBorn extends Enchantment implements IEnchantmentTooltip {
+    public AirBorn() {
+        super(Rarity.RARE, EnchantmentCategory.ARMOR_CHEST, new EquipmentSlot[]{EquipmentSlot.CHEST});
+    }
+
+    @Override
+    public int getMinCost(int level) {
+        return 2 + 10 * (level - 1);
+    }
+
+    @Override
+    public int getMaxCost(int level) {
+        return this.getMinCost(level) + 50;
+    }
+
+    public boolean checkCompatibility(Enchantment other) {
+        return !(other instanceof DiggingEnchantment) && super.checkCompatibility(other);
+    }
+
+    public static float getMiningSpeedMultiplier(LivingEntity entity, BlockState state) {
+        if (entity.onGround())
+            return 1f;
+        ItemStack chestStack = entity.getItemBySlot(EquipmentSlot.CHEST);
+        int lvl = chestStack.getEnchantmentLevel(NewEnchantmentsFeature.AIR_BORN.get());
+        if (lvl == 0)
+            return 1f;
+
+        return 5f;
+    }
+
+    @Override
+    public Component getTooltip(ItemStack stack, int lvl) {
+        return Component.translatable(this.getDescriptionId() + ".tooltip").withStyle(ChatFormatting.DARK_PURPLE);
+    }
+}