+++ changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Upcoming
+* New enchantments!
+  * Air Born: removes the mining penalty when not on ground 
 * Forge Hammers
   * Movement Speed -15% -> -10%
   * Entity Reach 1.5 -> 2